# Django 1.11.6 Upgrade and Optimisations 