--- a/WebApp/ISIS/autoreduce_webapp/autoreduce_webapp/settings.py.template
+++ b/WebApp/ISIS/autoreduce_webapp/autoreduce_webapp/settings.py.template
@@ -49,10 +49,21 @@ AUTHENTICATION_BACKENDS = (
 )
 LOGIN_URL = '/'
 
-TEMPLATE_PATH = os.path.join(BASE_DIR, 'templates')
-TEMPLATE_DIRS = (
-    TEMPLATE_PATH,
-)
+TEMPLATES = [
+    {
+        'BACKEND': 'django.template.backends.django.DjangoTemplates',
+        'DIRS': [os.path.join(BASE_DIR, 'templates')],
+        'APP_DIRS': True,
+        'OPTIONS': {
+                'context_processors': [
+                'django.template.context_processors.debug',
+                'django.template.context_processors.request',
+                'django.contrib.auth.context_processors.auth',
+                'django.contrib.messages.context_processors.messages',
+            ],
+        },
+    },
+]
 
 ROOT_URLCONF = 'autoreduce_webapp.urls'