--- a/WebApp/ISIS/autoreduce_webapp/autoreduce_webapp/settings.py.template
+++ b/WebApp/ISIS/autoreduce_webapp/autoreduce_webapp/settings.py.template
@@ -55,7 +55,7 @@ TEMPLATES = [
         'DIRS': [os.path.join(BASE_DIR, 'templates')],
         'APP_DIRS': True,
         'OPTIONS': {
-                'context_processors': [
+            'context_processors': [
                 'django.template.context_processors.debug',
                 'django.template.context_processors.request',
                 'django.contrib.auth.context_processors.auth',