--- a/WebApp/ISIS/autoreduce_webapp/autoreduce_webapp/settings.py.template
+++ b/WebApp/ISIS/autoreduce_webapp/autoreduce_webapp/settings.py.template
@@ -1,24 +1,23 @@
 import os
+
 BASE_DIR = os.path.dirname(os.path.dirname(__file__))
 
 # Quick-start development settings - unsuitable for production
 # See https://docs.djangoproject.com/en/dev/howto/deployment/checklist/
 
 # SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY = 'YOUR_SECRET_KEY'
+SECRET_KEY = 'YOUR-SECRET-KEY'
 
-# SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = True
-
-TEMPLATE_DEBUG = True
+# SECURITY WARNING: don't run with these turned on in production!
+DEBUG = False
+DEBUG_PROPAGATE_EXCEPTIONS = False
 
 ALLOWED_HOSTS = ['localhost', 'reducedev2.isis.cclrc.ac.uk']
-
-ADMINS = ()
+INTERNAL_IPS = ['localhost', '127.0.0.1']
 
 # Application definition
 
-INSTALLED_APPS = (
+INSTALLED_APPS = [
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
@@ -30,9 +29,10 @@ INSTALLED_APPS = (
     'reduction_viewer',
     'reduction_variables',
     'django_user_agents',
-)
+]
 
-MIDDLEWARE_CLASSES = (
+MIDDLEWARE_CLASSES = [
+
     'django.contrib.sessions.middleware.SessionMiddleware',
     'django.middleware.common.CommonMiddleware',
     'django.middleware.csrf.CsrfViewMiddleware',
@@ -41,12 +41,13 @@ MIDDLEWARE_CLASSES = (
     'django.contrib.messages.middleware.MessageMiddleware',
     'django_user_agents.middleware.UserAgentMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
-)
+]
 
-AUTHENTICATION_BACKENDS = (
+AUTHENTICATION_BACKENDS = [
     'autoreduce_webapp.backends.UOWSAuthenticationBackend',
     'django.contrib.auth.backends.ModelBackend',
-)
+]
+
 LOGIN_URL = '/'
 
 TEMPLATES = [
@@ -68,7 +69,6 @@ TEMPLATES = [
 ROOT_URLCONF = 'autoreduce_webapp.urls'
 
 WSGI_APPLICATION = 'autoreduce_webapp.wsgi.application'
-
 
 # Database
 # https://docs.djangoproject.com/en/dev/ref/settings/#databases
@@ -93,16 +93,14 @@ USE_I18N = True
 USE_L10N = True
 USE_TZ = True
 
-
 # Static files (CSS, JavaScript, Images)
 # https://docs.djangoproject.com/en/dev/howto/static-files/
 
-STATIC_ROOT = ''
 STATIC_URL = '/static/'
 STATIC_PATH = os.path.join(BASE_DIR, 'static')
-STATICFILES_DIRS = (
+STATICFILES_DIRS = [
     os.path.join('static'),
-)
+]
 
 # Logging
 # https://docs.python.org/2/howto/logging.html
@@ -147,59 +145,56 @@ LOGGING = {
     }
 }
 
-# ActiveMQ 
+# ActiveMQ
 
 ACTIVEMQ = {
-    'topics' : [
+    'topics': [
         '/queue/DataReady',
-        #'/queue/ReductionPending', - Only used by autoreduction server
         '/queue/ReductionStarted',
         '/queue/ReductionComplete',
         '/queue/ReductionError'
-        ],
-    'username' : 'YOUR-ACTIVEMQ-USERNAME',
-    'password' : 'YOUR-ACTIVEMQ-PASSWORD',
-    'broker' : [("YOUR-ACTIVE-MQ-HOST", 61613)],
-    'SSL' : False
+    ],
+    'username': 'YOUR-ACTIVEMQ-USERNAME',
+    'password': 'YOUR-ACTIVEMQ-PASSWORD',
+    'broker': [("YOUR-ACTIVE-MQ-HOST", 61613)],
+    'SSL': False
 }
 
 # File Locations
 
-#ARCHIVE_BASE = ''
 if os.name == 'nt':
-    REDUCTION_DIRECTORY = r'\\isis\inst$\NDX%s\user\scripts\autoreduction' # %(instrument)
-    ARCHIVE_DIRECTORY = r'\\isis\inst$\NDX%s\Instrument\data\cycle_%s\autoreduced\%s\%s' # %(instrument, cycle, experiment_number, run_number)
-    
+    REDUCTION_DIRECTORY = r'\\isis\inst$\NDX%s\user\scripts\autoreduction'  # %(instrument)
+    ARCHIVE_DIRECTORY = r'\\isis\inst$\NDX%s\Instrument\data\cycle_%s\autoreduced\%s\%s'  # %(instrument, cycle, experiment_number, run_number)
+
     TEST_REDUCTION_DIRECTORY = r'\\reducedev\isis\output\NDX%s\user\scripts\autoreduction'
     TEST_ARCHIVE_DIRECTORY = '\\isis\inst$\NDX%s\Instrument\data\cycle_%s\autoreduced\%s\%s'
 
 else:
-    REDUCTION_DIRECTORY = '/isis/NDX%s/user/scripts/autoreduction' # %(instrument)
-    ARCHIVE_DIRECTORY = '/isis/NDX%s/Instrument/data/cycle_%s/autoreduced/%s/%s' # %(instrument, cycle, experiment_number, run_number)
-    
-    TEST_REDUCTION_DIRECTORY ='/reducedev/isis/output/NDX%s/user/scripts/autoreduction'
+    REDUCTION_DIRECTORY = '/isis/NDX%s/user/scripts/autoreduction'  # %(instrument)
+    ARCHIVE_DIRECTORY = '/isis/NDX%s/Instrument/data/cycle_%s/autoreduced/%s/%s'  # %(instrument, cycle, experiment_number, run_number)
+
+    TEST_REDUCTION_DIRECTORY = '/reducedev/isis/output/NDX%s/user/scripts/autoreduction'
     TEST_ARCHIVE_DIRECTORY = '/isis/NDX%s/Instrument/data/cycle_%s/autoreduced/%s/%s'
 
-# ICAT 
+# ICAT
 
 ICAT = {
-    'AUTH' : 'YOUR-ICAT-AUTH',
-    'URL' : 'YOUR-ICAT-WSDL',
-    'USER' : 'YOUR-ICAT-USERNAME',
-    'PASSWORD' : 'YOUR-ICAT-PASSWORD'
+    'AUTH': 'YOUR-ICAT-AUTH',
+    'URL': 'YOUR-ICAT-WSDL',
+    'USER': 'YOUR-ICAT-USERNAME',
+    'PASSWORD': 'YOUR-ICAT-PASSWORD'
 }
 
 # Outdated Browsers
 
 OUTDATED_BROWSERS = {
-    'IE' : 9,
+    'IE': 9,
 }
 
 # UserOffice WebService
 
 UOWS_URL = 'https://fitbaweb1.isis.cclrc.ac.uk:8443/UserOfficeWebService/UserOfficeWebService?wsdl'
 UOWS_LOGIN_URL = 'https://users.facilities.rl.ac.uk/auth/?service=http://reduce.isis.cclrc.ac.uk&redirecturl='
-
 
 # Email for notifications
 
@@ -214,9 +209,9 @@ CERTIFICATE_LOCATION = 'C:\\certificates\\FITBAWEB1isiscclrcacuk.crt'
 # Constant vars
 
 FACILITY = "ISIS"
-PRELOAD_RUNS_UNDER = 100 # If the index run list has fewer than this many runs to show the user, preload them all.
-CACHE_LIFETIME = 3600 # Objects in ICATCache live this many seconds when ICAT is available to update them.
-USER_ACCESS_CHECKS = True # Should the webapp prevent users from accessing runs/instruments they're not allowed to?
-DEVELOPMENT_MODE = True # If the installation is in a development environment, set this variable to True so that
-                        # we are not constrained by having to log in through the user office. This will authenticate
-                        # anyone visiting the site as a super user
+PRELOAD_RUNS_UNDER = 100  # If the index run list has fewer than this many runs to show the user, preload them all.
+CACHE_LIFETIME = 3600  # Objects in ICATCache live this many seconds when ICAT is available to update them.
+USER_ACCESS_CHECKS = False  # Should the webapp prevent users from accessing runs/instruments they're not allowed to?
+DEVELOPMENT_MODE = True  # If the installation is in a development environment, set this variable to True so that
+                         # we are not constrained by having to log in through the user office. This will authenticate
+                         # anyone visiting the site as a super user