… WebApp/ISIS/autoreduce_webapp/autoreduce_webapp/utils.py