This pull request is an upgrade to the WebApp that allows it to work with the latest version of Django (1.11.6). This version is the last Python 2.7 long term support release, but will receive updates until at least [April 2020. ](https://www.djangoproject.com/download/#supported-versions)


This pull request also makes a few small changes to the WebApp:
- Most importantly it updates `reduction_viewer/views.py` lines 265-274 which make the loading of the run table a lot quicker. Running the WebApp on the same database with 390 runs, worst case it would take 7.5s to load, typically 4.0s. With these changes it takes worst case 1.5s, typically 300ms which is much more acceptable.
- Adds 'RB' infront of the Experiment Number to make it clearer it's an experiment number.
- Change the the size of the columns in the run table to make it clearer for smaller displays such as phones.

This can be seen running on [http://reducedev2.isis.cclrc.ac.uk:8080](http://reducedev2.isis.cclrc.ac.uk:8080)