@Anders-Markvardsen 2017-10-26
Also in setting.py.template line 11 should DEBUG not be set to FALSE by default to avoid the human mistake of forgetting to change this parameter on production? Unless you argue otherwise could you change this with this pull request also